requirements.txt

git+https://github.com/pytgcalls/pytgcalls.git
aiofiles>=23.2.1
aiohttp>=3.9.5
apscheduler>=3.10.4
asyncio>=3.4.3
bard>=0.2.0
beautifulsoup4>=4.12.3
cloudscraper>=1.2.69
daxxhub>=1.0.0
deepai>=1.1.0
dnspython>=2.4.2
emojis>=0.7.0
Faker>=25.2.0
ffmpeg-python>=0.2.0
future==0.18.3
git+https://github.com/AsmSafone/SafoneAPI.git
gitpython>=3.1.43
gpytranslate>=1.2.2
gtts>=2.4.0
hachoir>=3.1.3
heroku3>=5.2.0
httpx==0.25.2
instaloader>=4.9
img2pdf>=0.4.4
lexica-api==1.5.1
motor>=3.3.2
numpy>=1.26.4
openai>=1.0.0
opencv-python>=4.9.0.80
pickledb>=0.9.2
pillow==9.5.0
psutil>=5.9.8
pycountry>=23.12.12
pycryptodome>=3.20.0
pydub>=0.25.1
pyfiglet>=0.8.post1
pykeyboard>=0.2.4
pyshorteners>=1.0.1
pyrofork>=1.0.0
python-dotenv>=1.0.1
python-telegram-bot==20.7
python-whois>=0.7.3
pytz>=2024.1
pyyaml>=6.0.1
qrcode>=7.4.2
requests>=2.32.3
search_engine_parser>=0.6.3
speedtest-cli>=2.1.3
spotipy>=2.23.0
SpeechRecognition>=3.10.0
telegraph>=0.6.0
telethon>=1.34.0
tgcrypto>=1.2.5
unidecode>=1.3.8
uvloop>=0.19.0
wget>=3.2
youtube-search>=2.1.2
youtube-search-python>=1.6.6
yt-dlp>=2025.6.25
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
